+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/dom/typedarray/TypedArrays.java
@@ -205,7 +205,7 @@ public class TypedArrays {
 
     @JSBody(params = {"buffer"}, script = "" +
             "return buffer;")
-    private static native ArrayBufferViewWrapper getTypedArray(@JSByRef() byte[] buffer);
+    public static native ArrayBufferViewWrapper getTypedArray(@JSByRef() byte[] buffer);
 
     @JSBody(params = {"buffer"}, script = "" +
             "return buffer;")