+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/filesystem/FileDB.java
@@ -1,12 +1,9 @@
 package com.github.xpenatan.gdx.backends.teavm.filesystem;
 
-import com.badlogic.gdx.Files;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.files.FileHandle;
 import com.badlogic.gdx.utils.GdxRuntimeException;
-import com.github.xpenatan.gdx.backends.teavm.TeaApplication;
 import com.github.xpenatan.gdx.backends.teavm.TeaFileHandle;
-import com.github.xpenatan.gdx.backends.teavm.preloader.Preloader;
 import java.io.ByteArrayOutputStream;
 import java.io.FileFilter;
 import java.io.FilenameFilter;
@@ -22,22 +19,6 @@ import java.util.List;
  * @author noblemaster
  */
 public abstract class FileDB {
-
-    /**
-     * The singleton instance as we only have one file system DB.
-     */
-    private static FileDB INSTANCE = null;
-
-    protected FileDB() {
-        // hiding the constructor
-    }
-
-    public static FileDB getInstance() {
-        if(INSTANCE == null) {
-            INSTANCE = new FileDBManager();
-        }
-        return INSTANCE;
-    }
 
     public abstract InputStream read(TeaFileHandle file);
 
@@ -83,18 +64,13 @@ public abstract class FileDB {
         String[] paths = paths(file);
         FileHandle[] files = new TeaFileHandle[paths.length];
 
-        Preloader preloader = null;
-        if(Gdx.app != null) {
-            if(Gdx.app instanceof TeaApplication) {
-                preloader = ((TeaApplication)Gdx.app).getPreloader();
-            }
-        }
         for(int i = 0; i < paths.length; i++) {
             String path = paths[i];
             if((path.length() > 0) && (path.charAt(path.length() - 1) == '/')) {
                 path = path.substring(0, path.length() - 1);
             }
-            files[i] = new TeaFileHandle(preloader, this, path, file.type());
+
+            files[i] = Gdx.files.getFileHandle(path, file.type());
         }
         return files;
     }
@@ -138,6 +114,4 @@ public abstract class FileDB {
     public abstract long length(TeaFileHandle file);
 
     public abstract void rename(TeaFileHandle source, TeaFileHandle target);
-
-    public abstract String getPath();
 }