+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/filesystem/MemoryFileStorage.java
@@ -204,11 +204,6 @@ public class MemoryFileStorage extends FileDB {
         }
     }
 
-    @Override
-    public String getPath() {
-        return "";
-    }
-
     public String debugAllFiles() {
         String text = "";
         text += println("####### START DEBUG FILE: " + fileMap.size + "\n");