+++ CHANGES.md
@@ -1,9 +1,10 @@
 [1.0.0-SNAPSHOT]
 - Update teavm to 0.10.0
 - Pixmap now use Gdx2DPixmap
-- Add useNewFileHandle set to true as default. Gdx.files.internal/classpath use new class to hold files and add IndexedDB support for Gdx.files.local  
+- New Gdx.files.internal, classpath and local implementation. Local files uses IndexedDB.  
 - Call dispose when browser closes
 - Improve clipboard text copy/paste.
+- Change default sound/music api to howler.js
 
 [1.0.0-b9]
 - add TeaClassFilter printAllowedClasses() and printExcludedClasses()